NetSync: Fix size_t errors (revert to int) to fix netsync

+++ src/ds/data/data_buffer.cpp
@@ -5,16 +5,16 @@
 
 namespace ds {
 
-DataBuffer::DataBuffer(size_t initialStreamSize)
+DataBuffer::DataBuffer(unsigned initialStreamSize)
 	: mStream(initialStreamSize)
 {
 }
 
-size_t DataBuffer::size(){
-	size_t currentPosition = mStream.getReadPosition();
+unsigned DataBuffer::size(){
+	unsigned currentPosition = mStream.getReadPosition();
 
 	mStream.setReadPosition(ReadWriteBuffer::End);
-	size_t length = mStream.getReadPosition();
+	unsigned length = mStream.getReadPosition();
 	mStream.setReadPosition(currentPosition);
 
 	return length;
@@ -29,15 +29,15 @@ void DataBuffer::clear(){
 	mStream.clear();
 }
 
-void DataBuffer::addRaw(const char *b, size_t size){
+void DataBuffer::addRaw(const char *b, unsigned size){
 	mStream.write(b, size);
 }
 
-bool DataBuffer::readRaw(char *b, size_t size){
-	size_t currentPosition = mStream.getReadPosition();
+bool DataBuffer::readRaw(char *b, unsigned size){
+	unsigned currentPosition = mStream.getReadPosition();
 
 	mStream.setReadPosition(ReadWriteBuffer::End);
-	size_t length = mStream.getReadPosition();
+	unsigned length = mStream.getReadPosition();
 	mStream.setReadPosition(currentPosition);
 
 	if(size > (length - currentPosition))
@@ -47,7 +47,7 @@ bool DataBuffer::readRaw(char *b, size_t size){
 	return true;
 }
 
-void DataBuffer::add(const char *b, size_t size){
+void DataBuffer::add(const char *b, unsigned size){
 	add(size);
 	mStream.write(b, size);
 }
@@ -60,17 +60,17 @@ void DataBuffer::add(const wchar_t *cs){
 	add<std::wstring>(cs);
 }
 
-bool DataBuffer::read(char *b, size_t size){
-	size_t wsize = read<size_t>();
+bool DataBuffer::read(char *b, unsigned size){
+	unsigned wsize = read<unsigned>();
 	if(wsize != size) {
 		add(wsize);
 		return false;
 	}
 
-	size_t currentPosition = mStream.getReadPosition();
+	unsigned currentPosition = mStream.getReadPosition();
 
 	mStream.setReadPosition(ReadWriteBuffer::End);
-	size_t length = mStream.getReadPosition();
+	unsigned length = mStream.getReadPosition();
 	mStream.setReadPosition(currentPosition);
 
 	if(size > (length - currentPosition))
@@ -85,14 +85,14 @@ bool DataBuffer::read(char *b, size_t size){
 // Template specializations
 template <>
 void DataBuffer::add<std::string>(const std::string &s){
-	size_t size = (unsigned)s.size();
+	unsigned size = (unsigned)s.size();
 	add(size);
 	mStream.write(s.c_str(), size);
 }
 
 template <>
 void DataBuffer::add<std::wstring>(const std::wstring &ws){
-	size_t size = (ws.size())*sizeof(wchar_t);
+	unsigned size = (ws.size())*sizeof(wchar_t);
 	add(size);
 	mStream.write((const char *)(ws.c_str()), size);
 }
